+++ src/js/visual/csr/utils.js
@@ -25,6 +25,27 @@ ripe.CsrUtils.toPrecision = function(value, precision = 6) {
 };
 
 /**
+ * Normalizes a THREE.Object3D rotation by setting its axis only with positive
+ * values ranging from 0 to 2*PI.
+ *
+ * @param {THREE.Object3D} object3D An instance of a THREE.Object3D.
+ */
+ripe.CsrUtils.normalizeRotations = function(object3D) {
+    if (!object3D) return;
+
+    const range = Math.PI * 2;
+    let x = object3D.rotation.x % range;
+    let y = object3D.rotation.y % range;
+    let z = object3D.rotation.z % range;
+    x = x < 0 ? x + range : x;
+    y = y < 0 ? y + range : y;
+    z = z < 0 ? z + range : z;
+    object3D.rotation.x = this.toPrecision(x, 7);
+    object3D.rotation.y = this.toPrecision(y, 7);
+    object3D.rotation.z = this.toPrecision(z, 7);
+};
+
+/**
  * Calculates the shortest path in radians needed to rotate between two
  * points in a unit circle.
  *